Removed the features folder and moved contents up a level.
Views are not features, and so the features folder was unnecessary. Setting up
the project by Feature implies that features are somehow a logical grouping
of components, views, reducers, action creators and actions. But we just
weren't there yet.

We should revisit this in the future, though. It's probably worth making
the change at some point, in order to make the dev workflow easier.