2020 09 01 monarch review and fixes (#3)
* Removed all references to launches

* Removed the features folder and moved contents up a level.

Views are not features, and so the features folder was unnecessary. Setting up
the project by Feature implies that features are somehow a logical grouping
of components, views, reducers, action creators and actions. But we just
weren't there yet.

We should revisit this in the future, though. It's probably worth making
the change at some point, in order to make the dev workflow easier.

* Cleaned up Home.tsx

* Started a new view layer

* Added a simple Button and Usercard

* Remove build/ and update gitignore

* Removed React.SFC in favour of React.FC since React.SFC is deprecated

Co-authored-by: Monarch Wadia <monarch@mintbean.io>
Co-authored-by: Jimmy Peng <jimmy.peng@mintbean.io>