Change filter for upcoming meets
Previously, meets currently taking place would not be categorized as
upcomingMeets nor pastMeets, meaning they would not be displayed.
This is because the filter for upcomingMeets checks if startTime is in
the past, while the filter for pastMeets checks if endTime is in the
past.

I have changed the upcomingMeets filter to endTime so current meets
is now categorized as upcoming for now.