Fix the clipping of modals
Previously, if a modal was long, like the "edit meet" modal, and
originated near the bottom of the screen, the vertical scroll would not
allow you to scroll down to view the entire modal, leaving the modal
clipped and in cases of modal forms, unable to be submitted.

This issue was caused by the addition of the `overflow-hidden` class to
the global layout. While its original purpose was to prevent the
horizontal scrollbar from being rendered in some cases, it also
prevented vertical scroll in cases of screen overflow, such as a modal
expanding vertically while at the bottom of the screen.

I have thus (temporarily) fixed the problem by changing the class from
`overflow-hidden` to `overflow-x-hidden` to prevent only horizontal
scroll while leaving vertical scroll intact, allowing the entire modal
to be scrollable into view.

It is my opinion that using the overflow-hidden classes are not an
optimal solution to the phantom horizontal scrollbar issues, as it
fixes the problem superficially and does not address the root cause
of the unnecessary horizontal scrollbar appearances, which are likely
css layout issues.